+++ src/Enums/NodeGroup.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Vdhicts\Cyberfusion\ClusterApi\Enums;
+
+class NodeGroup
+{
+    public const ADMIN = 'Admin';
+    public const APACHE = 'Apache';
+    public const NGINX = 'nginx';
+    public const DOVECOT = 'Dovecot';
+    public const MARIADB = 'MariaDB';
+    public const POSTGRESQL = 'PostgreSQL';
+    public const MAIN = 'Main';
+    public const PHP = 'PHP';
+    public const BORG = 'Borg';
+    public const FAST_REDIRECT = 'Fast Redirect';
+}