+++ CHANGELOG.md
@@ -7,6 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 cluster API. See the changelog of the [cluster API](https://cluster-api.cyberfusion.nl/redoc#section/Changelog) for 
 detailed information.
 
+## [1.43.0]
+
+### Added
+
+- Add nodes endpoint.
+- Add `random_delay_max_seconds` attribute to the `Cron` model.
+
 ## [1.42.0]
 
 ### Added